wip: fix(B2): census, bridge, pointer register and tidy findings of the batch
- AssetReferenceCensus reads the MSAS subsystem (MeshSourceAsset.cpp in its premake) and counts a
  mesh source asset's material slots as references, resolved by the same HandleForGuid fold.
- ImportedMeshAsset goes through Geometry::Bridge (EditMeshFromMeshAssetData, SavedFormFromEditMesh)
  instead of including EditMesh headers.
- PointerOwnership: rows for VulkanGpuProfiler::m_Recording and GpuScopeRecorder::m_Stacks (RT2g),
  census 455 / 996.
- clang-tidy findings on the batch's lines fixed without suppressions.

+++ Desert/Tests/TestSupport/cooked_static_mesh.hpp
@@ -27,7 +27,7 @@ namespace Desert::TestSupport
     {
         static const uint32_t        base = std::random_device{}() % 1000000u;
         static std::atomic<uint32_t> counter{ 0 };
-        const float                  width = 1.0f + static_cast<float>( base ) + 1000000.0f * counter++;
+        const float width = 1.0f + static_cast<float>( base ) + 1000000.0f * static_cast<float>( counter++ );
 
         Assets::MeshSourceAsset asset;
         asset.Guid              = Common::Content::AssetGuid::Generate();
